feat: move cp and eff surfaces to a common chart
